SAVE-1/SND-1: 4-byte binary fields written or read as 8-byte long
SAVE-1 (campaign save round-trip). Four campaign save files write their size
prefix with sizeof(long) while the loader reads int32_t, so nothing the Linux
build saves can be read back:

  SaveBaseObjectives  size, newsize  <- LoadBaseObjectives  reads int32_t
  SaveObjectiveDeltas csize          <- LoadObjectiveDeltas reads int32_t
  SaveUnits           size           <- LoadUnits           reads int32_t

Also four SaveSize() calculations still counted 8 bytes for fields whose
writers were already corrected to 4 (flight fuel_burnt, package package_flags,
waypoint Flags, objective obj_flags) -- each over-reserved by 4 bytes per record.

SND-1 (RIFF parsers). CSoundMgr::LoadRiffFormat read the RIFF size field as
long. At 8 bytes that read consumes the size field AND the 'WAVE' tag, so the
strncmp(buffer, "WAVE", 4) immediately below it compared against 'fmt ' and the
function returned 0 for every file. Verified against real game WAVs: old width
returns 0 for BoomA1/biggun1/LSpdTone, new width returns 115200/152320/45072.
Its caller F4StartStream ignores the return and uses the WAVEFORMATEX anyway,
so every streamed sound was built from an uninitialized stack header.

Same defect in FillRiffInfo's chunk walk (read 8 bytes, advance 4) and in both
SkipRiffHeader overloads -- those two are currently unreferenced, but they feed
a garbage size straight into fread(buffer, size, 1, fp) over a char[256], so
they were a stack smash waiting for a caller.

Verified TE 2 reaches 3D (63 samples) with zero crashes.

+++ src/campaign/camplib/campwp.cpp
@@ -220,7 +220,7 @@ int WayPointClass::SaveSize(void)
             + sizeof(uchar)
             + sizeof(uchar)
             + sizeof(uchar)
-            + sizeof(ulong);
+            + sizeof(uint32_t);   // FF_LINUX: Flags is written as uint32_t
     return size;
 }
